Remove recent search or starred word from side widget in weird case
The case in which we wish to do this is when the user has filtered
for all the subjects of the current profile and then updates the
subjects of a word which happens to belong in the recent searches or
starred words. If that word used to belong to a single subject and
after the update it no longer belongs to it, then the corresponding
recent search or starred word should be removed.